--- a/src/desktop/apps/artist/server.tsx
+++ b/src/desktop/apps/artist/server.tsx
@@ -58,6 +58,8 @@ app.get(
         },
       })
 
+      res.locals.PAGE_CACHE = { status, key: req.url, html: layout }
+
       res.status(status).send(layout)
     } catch (error) {
       next(error)