--- a/src/desktop/apps/experimental-app-shell/server.tsx
+++ b/src/desktop/apps/experimental-app-shell/server.tsx
@@ -82,6 +82,7 @@ app.get(
         },
       })
 
+      res.locals.PAGE_CACHE = { status, key: req.url, html: layout }
       res.status(status).send(layout)
     } catch (error) {
       console.log(error)