--- a/src/desktop/components/split_test/running_tests.coffee
+++ b/src/desktop/components/split_test/running_tests.coffee
@@ -24,11 +24,4 @@
 # this should export empty Object
 # module.exports = {}
 
-module.exports = {
-  client_navigation_v2:
-    key: "client_navigation_v2"
-    outcomes:
-      control: 50
-      experiment: 50
-    edge: "experiment"
-}
+module.exports = {}