cc @damassi @zephraph 

This implements a [relatively] simple mechanism for caching the full HTML of pages.

Which 'types' of pages to cache (artist, artwork, collection, etc.), and how long to cache are all configurable via ENV (as-is this entire feature, which can be enabled with a feature flag).

Additionally, there are some cache namespace and version ENV vars which can be used to effectively bust the cache anytime (for instance, after a deploy where a change is desired to be visible right away).

It's pretty simple, and locally seems to work great. Here are some local (not particularly realistic) benchmarks:

Without cache:

```
Document Path:          /artist/andy-warhol
Document Length:        274180 bytes

Concurrency Level:      1
Time taken for tests:   4.894 seconds
Complete requests:      5
Failed requests:        4
   (Connect: 0, Receive: 0, Length: 4, Exceptions: 0)
Total transferred:      1389383 bytes
HTML transferred:       1385812 bytes
Requests per second:    1.02 [#/sec] (mean)
Time per request:       978.766 [ms] (mean)
Time per request:       978.766 [ms] (mean, across all concurrent requests)
Transfer rate:          277.25 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:   781  979 190.5   1048    1223
Waiting:      781  978 190.4   1048    1223
Total:        781  979 190.5   1048    1224

Percentage of the requests served within a certain time (ms)
  50%   1020
  66%   1077
  75%   1077
  80%   1224
  90%   1224
  95%   1224
  98%   1224
  99%   1224
 100%   1224 (longest request)
```

With cache:

```
Document Path:          /artist/andy-warhol
Document Length:        279504 bytes

Concurrency Level:      1
Time taken for tests:   0.752 seconds
Complete requests:      5
Failed requests:        0
Total transferred:      1401091 bytes
HTML transferred:       1397520 bytes
Requests per second:    6.65 [#/sec] (mean)
Time per request:       150.388 [ms] (mean)
Time per request:       150.388 [ms] (mean, across all concurrent requests)
Transfer rate:          1819.63 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:   110  150  27.0    156     184
Waiting:      109  149  25.6    156     179
Total:        110  150  27.0    156     184

Percentage of the requests served within a certain time (ms)
  50%    152
  66%    160
  75%    160
  80%    184
  90%    184
  95%    184
  98%    184
  99%    184
 100%    184 (longest request)
```

It can be safely merged (since it's opt-in and fully disabled by default), and we can experiment on staging (and prod?!).

I'm curious to hear other's thoughts. It's not as 'complete' a solution as true edge caching with a CDN, but there are a host of hurdles to overcome for that, whereas this winds up being easier to understand and immediately implement, and possibly gives us a lot of benefit.