@mzikherman 2020-02-13
Great questions @joeyAghion.

> How does this behave across deploys? If the entire response is cached including asset references, that means the old code version will essentially continue to be in use until all the cached pages expire. "Fixes" would take that long to be fully reflected, unless we proactively bumped the page cache version config. This might be fine, but do you think it's worth adding the code version to the cache key to avoid this?

Sounds like you're thinking about updating the version, effectively, on each deploy. That makes sense to me! Especially with the cached references to old client-side assets.

---

> How sure are we that pages' HTML doesn't include anything that is time or session-sensitive? I'm thinking of things like timers, A/B test assignments, and the like.

> Do user agents need to be considered in deciding when to serve cached content, and what responses not to cache? I recall there was a code change recently to serve just one breakpoint to Googlebot. That might interfere with this kind of caching.

To me, both of these questions are related. They bring up possible request-specific data and scenarios, which naturally break with full caching.

I think we need to understand the limitations, and be able to live with, or dynamically turn on/off, this kind of thing.

For Google specifically, it sounds like Googlebot user agents should effectively skip this middleware entirely? That's easy enough (although, should we inversely cache and serve them their special version?), to essentially customize, or opt-in/out something for Googlebot specifically.

For AB tests, the 'is mobile' or 'is eigen' flag (which might affect some layout or CSS), they very well might run into unexpected issues. So we can include those in the cache key?

For user agents generally, I think since our Reaction apps ship the full payload and rely on the browser to immediately be able to render the correct one based on actual size, I think we're actually safe to do this. The 'is mobile' or 'is eigen' flags may control....something, tho.