@damassi 2020-02-13
My one thought is that should this be merged, we now have two layers of cache operating (albeit different kinds) -- a cache at the Relay network layer on the client, and a cache on the server. It might be better to consolidate this at a single point, namely at `buildServerApp`, located [here](https://github.com/artsy/reaction/blob/master/src/Artsy/Router/buildServerApp.tsx#L58). 

Also, it might be worth exploring handling cache server-side via `relay-network-modern`, similar to how we're handling it on the client:https://github.com/artsy/reaction/blob/master/src/Artsy/Relay/createRelaySSREnvironment.ts#L98-L106