@mzikherman 2020-02-13
@damassi I intentionally don't want to limit this to Relay stuff. This is intended to be general Force middleware that might work on the homepage, partner profile page, etc. So overall definitely thinking to keep this in Force and not Reaction, for instance.

In terms of the Relay client cache...that's really only in play on /collect suite (and hypothetically if we can get client side routing fully rolled out, other apps), right? Since this is just the cache for an initial SSR call, I don't think that those will wind up playing poorly together. In fact, imagine the possibilities of client-side routing plus page caching! 😍