@damassi 2020-02-14
As long as stale asset issues and other edge-cases can be ironed out, that seems like it would be super fast indeed 😄