@zephraph 2020-02-14
Could I suggest that we drastically reduce the cache time? What % of workload to force would we reduce if we just cached for 5 or 10 seconds. I suspect it’d have a measurable impact without raising as many staleness concerns.

At Scripps FoodNetwork and HGTV had a 10 second TTL between the origin and edge cache. That offloaded a significant portion of traffic. Here we’re still dealing with traffic but hoping to benefit from reduced CPU load. Can we play with those numbers a bit?

Further, is there another layer in our stack that we could implement caching in front of force?