@mzikherman 2020-02-14
> Could I suggest that we drastically reduce the cache time? What % of workload to force would we reduce if we just cached for 5 or 10 seconds. I suspect it’d have a measurable impact without raising as many staleness concerns.

Yup, I'll update the config to be in seconds (vs. minutes), and maybe set it to...10 as default?

> Further, is there another layer in our stack that we could implement caching in front of force?

Yup, a proper CDN for one 😆 . That has a number of issues (in addition to some of the ones present in this PR). Another thought would be some sort of snapshotting or pre-rendering setup? That feels iffy though.  The approach in this PR felt a little bit more...balanced, if that makes sense (besides the fact we're caching the whole world!).