@damassi 2020-02-14
Another thought (and this applies only to our new pages) is to only cache html returned from [`buildServerApp`, versus the output from stitch](https://github.com/artsy/force/blob/master/src/desktop/apps/artist/server.tsx#L51). That way we don't have to worry about the shell.