@mzikherman 2020-02-14
Did a little refactor that cleans up the middleware (and route handlers that were invoking the cache setting). Also updated the cache time to be 30 seconds, configurable.

Outstanding issues:
  - **Googlebot**: what should we do for these requests? Safe to just exclude from this completely for now? cc @joeyAghion do you have that PR handy where you mentioned some specific CSS or layout thing?

  - **Other request-specific data**: AB tests are definitely a thing to deal with. Since these can affect the SSR-rendered content, I think we need to include currently running AB tests in the cache key (not a _huge_ deal tbh). What about other request-specific data? I think user agent might be ok, (for modern Reaction apps), for the reasons mentioned earlier. Other stuff?